refactor(engine): S2 Stage 4 — extract Waveform mipmap bindings to XlethSvcWaveform TU
Verbatim move of the four waveform_* handlers (Waveform_GetRegionPeaks,
Waveform_GetRawSamples, Waveform_GetClipPeaks, Waveform_GetFilePeaks) out of
XlethEngineService.cpp into engine/src/service/XlethSvcWaveform.cpp/.h, along
with wfbLog/WFB_LOG (#ifdef XLETH_DEBUG debug logging) and the static
applyPeakTransforms helper — every caller of these three lives inside the
moved block. The new TU is self-contained on the xleth::svc globals
(g_timeline, audioEngine, sampleBank, g_mipmapCache) extracted in Stage 1; it
needs none of XlethSvcShared.h's isInitialised()/BridgeCallLog. dispatch()
in XlethEngineService.cpp is untouched — only addition is the
XlethSvcWaveform.h include, mirroring the Stage 2 MIDI pattern.

Zero behavior change: ctest 43/47, identical to the pre-move baseline (same
4 known failures: test_flip_orientation_golden, test_effects, test_reverb,
test_real_render). Bridge contract 12/15, identical baseline. vitest 2504
passed with the same 2 pre-existing local-only mixer/FxBadge failures. App
smoke 3/3. Live end-to-end verified against the rebuilt addon: all four
handlers export correctly and waveform_getFilePeaks generates a mipmap and
returns non-trivial peak data from a real WAV file, including the
cache-hit path on a second call.

See docs/S2_SPLIT_PLAN.md §4 Stage 4.
